Prevent incorrect use of `flatMap`
Swift will happily coerce functions of the type `A -> B` into functions
of the type `A -> B?`. That means that using `flatMap` with a function
that doesn't return an optional will compile. But then at runtime, you
will run into weird behaviour that you don't expect, or worse, Swift
will segfault.

In order to prevent these problems from happening in the future, we're
going to throw errors ourselves if `flatMap` is used incorrectly with a
message letting the user know that the type of the function isn't
correct. Using a default (but ultimately unused) implementation of `map`
will push people in the right direction if they go digging.

Credit to @mattrubin for the idea:
https://github.com/thoughtbot/Argo/issues/34#issuecomment-70166588